Object_Tracking/TrackTacular/requirements.txt
kornia==0.6.11
matplotlib==3.5.2
opencv-python==4.7.0.72
lightning[extra]==2.0.2
scipy==1.10.1
einops==0.6.1
nuscenes-devkit==1.1.10
numpy==1.24.2
lap==0.4.0
motmetrics==1.4.0
Pillow==9.5.0
pytorch_metric_learning==2.2.0
efficientnet_pytorch
timm==0.9.2
imgaug==0.4.0